\name{autoGate-package}
\alias{autoGate-package}
\alias{autoGate}
\docType{package}
\title{
  \packageTitle{autoGate}
}
\description{
  \packageDescription{autoGate}
}
\details{
}
\author{
  \packageAuthor{autoGate}

  Maintainer: \packageMaintainer{autoGate}
}
\references{
}

\keyword{ package }
\seealso{
  \code{\link[autoGate]{trim.fcs}}
}
\examples{
}
