% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFlowFrames.R
\name{trim.fcs}
\alias{trim.fcs}
\title{Trim .fcs files to remove debris and doublets.}
\usage{
trim.fcs(dir_path, pattern = "*.fcs", do_plot = F)
}
\arguments{
\item{dir_path}{a directory path containing the .fcs files  to be parsed or folders to be recursed through.}

\item{pattern}{a regex pattern to match particular .fcs files. Default is \code{"*.fcs"} matching all .fcs files.}

\item{do_plot}{a Boolean flag to determine whether to produce plots showing the trimming of each flowFrame. Defaults to \code{FALSE}.}
}
\description{
\code{trim.fcs} uses mixture models to cluster bacteria from background debris and fits a linear model to SSC-H vs SSC-A to remove doublets.
}

